refactor: extract result dialog helpers
